Add --moduleBase command-line option.
With the --moduleBase option, a user can change the interpretation
of absolute paths that begin with a single '/' in a triple-slash
reference path or include() path.

If the --moduleBase option is used, absolute paths are prepended
with the path provided as an option to --moduleBase.

For example, when run with "--moduleBase /foo/bar", the following
reference path:

   ///<reference path="/quux/test.ts"/>

becomes "/foo/bar/quux/test.ts".